stream-csv-as-json: split modules, followed lint suggestions.
