+++ types/stream-csv-as-json/tsconfig.json
@@ -9,7 +9,8 @@
         "typeRoots": ["../"],
         "types": [],
         "noEmit": true,
-        "forceConsistentCasingInFileNames": true
+        "forceConsistentCasingInFileNames": true,
+        "strictFunctionTypes": true
     },
     "files": ["index.d.ts", "stream-csv-as-json-tests.ts"]
 }
